Structure backend API to controllers
