+++ backend/api/serverless.ts
@@ -10,7 +10,7 @@ const app = Fastify({
 });
 
 // Register your application as a normal plugin.
-app.register(import("../src/server"), {
+app.register(import("../src/server-router"), {
     prefix: '/'
 });
