+++ frontend/app/components/appliances-forms/HeatPumpWaterHeaterForm.tsx
@@ -26,7 +26,7 @@ const HeatPumpWaterHeaterForm = () => {
   const [firstHourRating, setFirstHourRating] = useState("40");
 
   const fetchData = async () => {
-    const apiUrl = `https://electric-machines-h6x1.vercel.app/api/v1/appliance?applianceType=hpwh&capacityMin=1&capacityMax=1000&uef=1&fhr=1`;
+    const apiUrl = `https://electric-machines-h6x1.vercel.app/api/v1/appliance/appliance?applianceType=hpwh&capacityMin=1&capacityMax=1000&uef=1&fhr=1`;
     const response = await fetch(apiUrl);
     const data = await response.json();
     setResults(data);
@@ -37,7 +37,7 @@ const HeatPumpWaterHeaterForm = () => {
   }, []); // runs once
 
   // sample API call:
-  // https://electric-machines-h6x1.vercel.app/api/v1/appliance?applianceType=hpwh&tankCapacityGallons=40&uniformEnergyFactor=2.5&firstHourRating=60
+  // https://electric-machines-h6x1.vercel.app/api/v1/appliance/appliance?applianceType=hpwh&tankCapacityGallons=40&uniformEnergyFactor=2.5&firstHourRating=60
 
   // Sample API response
   // [
@@ -59,7 +59,7 @@ const HeatPumpWaterHeaterForm = () => {
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
-    const apiUrl = `https://electric-machines-h6x1.vercel.app/api/v1/appliance?applianceType=hpwh&capacityMin=${tankCapacityGallons}&capacityMax=${
+    const apiUrl = `https://electric-machines-h6x1.vercel.app/api/v1/appliance/appliance?applianceType=hpwh&capacityMin=${tankCapacityGallons}&capacityMax=${
       tankCapacityGallons + 10
     }&uef=${uniformEnergyFactor}&fhr=${firstHourRating}`;
     // console.log(apiUrl);
